Avoid piling pending transport requests
Should the transport `process` not resolve within publish interval time,
avoiding piling another flush and `process` on top of it. This avoids
piling on requests in "pending" state during network instability or
transport unavailability.

The logs that are not successfully sent will be eventually be purged,
but this is already the case now. Instead of multiple concurrent pending
requests which might take and put logs back to buckets the publishes
will create one transporter request at a time which will send all the
pending logs.