Fix incrementing with tags
CHAN-300

Previously, when value was not present in arguments when calling
the increment/decrement functions with tags set via withTags,
the additional tags were sent separately from the actual tags array,
resulting in custom tags not showing up in Datadog.

Fix this by setting value to increment by with 1 and tags with the
value parameter when value isn't an integer (is an array of tags).