Implement limit to amount of consecutive retries
An issue occurred where the same batch of logs was re-sent over and over
again. While the source of the issue remains unclear, it would appear
that the logs were re-added to the buckets. This is the intended
behaviour at the moment, but only in the case that the sending of the
logs fails.

To address such a case, this change introduces a limit to the number of
times a single batch of logs is retried. Upon reaching this limit, the
failed batch is no longer re-added to the buckets.

PRIS-13