Ignore already acknowledged errors
Before `client-logger-js` was created we were ignoring all errors marked
with `smAcknowledged` as it means that they are already logged at some
point.

The common case:

```
sm.require.observable(['some module'])
 .flatMap(doSomeStuff)
 .subscribe(moreStuff, sm.logger.error);
```

In this case requres observable already logs the error with appropriate
level, records it in metrics and marks it as `sm.require.observable`.
There is no need to log it second time with level error as this causes
a lot of error noise in sentry.

CHAN-962