Avoid converting null to {} while formatting
Sending empty object as a default value to backend is confusing and
causes issues for indexer. For example when a field is expected to be a
number, then `{}` is a mismatch and might be ignored by the indexer.

A clear case that should not be casted to `{}` is `null`. Literal null
should stay unchanged as that is a safe fallback for ES fields.